Backend/requirements.txt
#fastapi
#uvicorn
#chromadb
#sentence-transformers
#sqlalchemy
#aiosmtplib
#python-dotenv
#requests
#pydantic

aiosmtplib==4.0.0
chromadb==0.6.3
fastapi==0.115.12
pydantic==2.11.1
python-dotenv==1.1.0
requests==2.32.3
sentence-transformers==4.0.2
SQLAlchemy==2.0.40
uvicorn==0.34.0

uvicorn main:app --host 0.0.0.0 --port 8000 --reload